requirements.txt
streamlit>=1.28.0
pytesseract>=0.3.10
opencv-python>=4.8.0
numpy>=1.24.0
requests>=2.31.0
reportlab>=4.0.0
Pillow>=10.0.0
